\name{KbaseKit-package}
\alias{KbaseKit-package}
\alias{KbaseKit}
\docType{package}
\title{
Basic Toolkit for the Kbase MG-RAST Server
}
\description{
An object framework for comparative analysis of metagenomes; facilities to download public and user-owned data and metadata from the MG-RAST service; a function library mirroring existing browser-based analysis routines; additional analysis techniques implemented in tandem with ongoing research and development.
}
\details{
\tabular{ll}{
Package: \tab KbaseKit\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2012-01-05\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel Braithwaite
}
\references{
\emph{The metagenomics RAST server: a public resource for the automatic phylogenetic and functional analysis of metagenomes}, Meyer et al., 2008\cr
\href{http://metagenomics.anl.gov}{http://metagenomics.anl.gov}
}
\keyword{ package }
\keyword{ Kbase }
\keyword{ kbase }
\seealso{
\code{\link{kbGet}}
}
\examples{
## Forthcoming!
}
